sqs integration tests
port the s3 integration tests -> sqs integration tests

integration tests are "opt-in" and require that ENABLE_INTEGRATION_TESTS
environment variable is set, otherwise they will skip. This allows the
general unit tests to execute as expected with "go test ./..."

The integration tests require to be run with AWS IAM credenitals in the
environment (or iam-profile/metadata). You may also optionally accept a
PERMISSION_BOUNDARY_ARN environment variable which will ensure any IAM
users created will have the attached boundary (this is useful for
testing stricter permissions setups but not required).

The integration tests perform a basic flow throug the life cycle of a
queue:

* Provision
* Wait
* Bind
* Wait
* Update
* Wait
* Use credenitals from binding
* Unbind
* Wait
* Deprovision

since the sqs broker uses async bindings we must also remember to poll
for state after creating bindings too

we perform an update of a single simple parameter and wait for success,
as a sanity check that update works, testing the correct operation of
each of the options is out of scope of this test.

Co-Authored-By: Philip Potter <philip.potter@digital.cabinet-office.gov.uk>
Co-Authored-By: Alex Monk <alex.monk@digital.cabinet-office.gov.uk>