Queue cloudformation template
The provision / deprovision operations map to the creation / deleteion
of the queue cloudformation template. We use "goformation" (go-ified
types of cloudformation) to build the aws resources we shall create when
users request a queue.

The provision parameters map directly to the QueueParam struct, so their
json struct tags become the user interface for setting the more advanced
options.

We avoid the use of cloudformation's parameter system and opt for using
the go-types directly to avoid issues with cloudformation parameters
being stringly-typed.

Provisioning will actually generate two AWS SQS Queue resources, one
will be designated the "main" queue and one the "dead-letter" queue.

By default, no messages will be sent to the dead-letter queue, but by
setting the `RedriveMaxReceiveCount` parameter during `create-service`
or `update-service` you may enable activiating delivery of failed items
to the dead-letter queue. This "two queues by default" vastly simplifies
the common case where you want a dead-letter queue, without the need of
tightly coupling two cloudfoundry service instances together. (you only
pay for the use of the queue not the queue itself, so this does not have
an impact on cost)

The cloudformation outputs are used to return values to the code that
will later be used in bindings. Here we return the QueueURLs / ARNs so
that the provisioner code can make use of it.