iam user cloudformation template for binding
each bind/unbind operation maps to a create/destroy of the user
cloudformation template.

we use goformation here to allow us to programatically build
cloudformation templates with go types and avoid the brittle use of the
stringly-typed cloudformation input parameters.

On bind:

* an IAM User resource will be created under a given prefix/path
* an accesskey/secret generated
* a policy attached restricted to sqs actions and scoped to the given queue arns
* a permission boundary attached (if provided in config)
* tags are attached to make the resources easier to identify/recharge

Currently the accesskey and secret is returned in the cloudformation
outputs (for consumption by the broker code)... cloudformation outputs
are not neccasarily a well know place where sensitive values are kept
and as such it would be easy to misconfigure IAM policies with "read"
without thinking this is giving access to secrets. We intend to improve
this part by writing the key/secret to ssm/paramstore in follow up work.

Co-Authored-By: Philip Potter <philip.potter@digital.cabinet-office.gov.uk>
Co-Authored-By: Alex Monk <alex.monk@digital.cabinet-office.gov.uk>