update GetBinding to fetch credentials from ssm
Previously we had to fetch both the queue's stack and the binding's
stack and then assemble the outputs into the credentials.

This changes things so that we assume the entire credentials blob is
already computed and stored in ssm parameter store waiting for us to
fetch it and return it to the user.

We fetch the name of the parameter from the binding's stack's output
(since this value is not sensitive), then call GetParameter to extract
the value.

This simplifies the code a fair bit as we no longer need to fetch both
stacks, and it also means the provider can be a bit dumb about the
credentials type, the only assumption is that it's JSON encoded.

As part of this change I have also been updating all the names to be
consistant, calling the queue's Primary/Secondary instead of Main/DL as
it makes things much easier to read.