update bind template to write credentials to ssm
we want to write the credentials into parameter store so they can be
safely accessed/audited instead of being easily leaked in cloudformation
outputs.

The provider expects to be able to fetch the entire JSON credentials
blob, so that's what we want to write into the store.

To do this, we build a "template" making use of the cloudformation "Sub"
intrinsic function, where values like "${thing}" will get interpolated
as if they are calls fetching the `Ref` from the resource named "thing".

This lets us build the JSON blob using values that will only exist after
the cloudformation stack is provisioned.

Errr "what the hell is that test" I hear you cry ... yeah "that test",
the one where I decode the template and type cast every value about 10
times ... that is using goformation's ability to pretend to be
cloudformation and evaluating the intrinsic function so I can try and
check that the template will produce a parameter store value with a JSON
blob with the expected keys

The Credentials blob is now using snake_case_format too for consistancy
with the other GOV.UK PaaS broker credentials.

As part of this change, the continued migration of naming from Main ->
Primary and DL -> Secondary is mixed in (sorry). This renaming better
reflects the queues use and makes things easier to read (especially in
tests).