switching from ssm -> secretsmanager
cloudformation does not support SecureString parameter type for SSM ...
it feels wrong to store these sensitive values unencrypted, so we switch
out the PutParameter/GetParameter logic for
PutSecretValue/GetSecretValue.

The design is identical (writing the blob of credentials into the secret
store and fetching them out from GetBinding), but it uses secretsmanager
instead.