add support for fifo queues
fifo queues have a strange requirement that the queue names must end
with a ".fifo" suffix, since we did not explicitly test the fifo queue
in the integration tests this went unnoticed.

the integration tests have been extended to now run two passes of the
full lifecycle ... once for "standard" and once for "fifo" queues. Since
fifo queues require a MessageGroupID set in the SendMessage calls we
need to tweak the send payload per test.

Queue names are now generated by functions and conform to the following
pattern:

$PREFIX-$UUID-$KIND.$EXT

Where:

$PREFIX is something like "paas-sqs-broker"
$UUID is the service instance id
$KIND is the primary/secondary marker
$EXT is the extension for std/fifo queue