Allow specifying an additional policy on managed users
We want to be able to restrict IAM Users managed by the SQS broker to
a particular domain (eg VPC, source IP, whatever).  This adds a
configuration item `additional_user_policy` which is the ARN of a
managed policy which will be attached to every managed IAM User.  This
enables setting a "local-only" policy that denies all requests without
the appropriate value of `aws:SourceVpc`.  See README update for
details.