update the README with a more realistic example IAM policy
The `aws:SourceVpc` only works if you access AWS services over a VPC
Endpoint, which we don't do.  This changes the example to be
representative of something we believe will work.

It uses RFC5737 IP addresses reserved for documentation [1] to avoid
any issues with using real (private or public) IPs.

[1]: https://tools.ietf.org/html/rfc5737